refactor: 홈 화면 첫번째 섹션 컴포넌트 분리 및 css 수정
